+++ src/utils.ts
@@ -22,9 +22,9 @@ export class IPXError extends Error {
   statusMessage?: string
 }
 
-export function createError (message: string, statusCode: number): IPXError {
-  const err = new IPXError(message)
-  err.statusMessage = 'IPX: ' + message
+export function createError (statusMessage: string, statusCode: number, trace?: string): IPXError {
+  const err = new IPXError(statusMessage + (trace ? ` (${trace})` : ''))
+  err.statusMessage = 'IPX: ' + statusMessage
   err.statusCode = statusCode
   return err
 }